emoshape/requirements.txt
aioamqp==0.14.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-speech==1.3.2
googleapis-common-protos==1.52.0
grpcio==1.30.0
idna==2.10
pamqp==2.3.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.9
